fix: Moved constants server and app into src. Wow.
