+++ src/api/v1/controllers/trackController.ts
@@ -1,6 +1,6 @@
 import { Request, Response, NextFunction } from "express"
 import { TrackMetadata } from "../models/models"
-import { HTTP_STATUS } from "../constants/httpConstants"
+import { HTTP_STATUS } from "../../../constants/httpConstants"
 import { successResponse } from "../models/responseModel"
 import { createTrackService, deleteTrackService, getAllTracksService, getTrackByIdService, updateTrackService } from "../services/trackService"
